Move finish_connect into Tcp type
